\name{IncPeriod-internal}
\alias{IncPeriod-internal}
\alias{fun_F}
\alias{fnlik_general}
\title{The main function of the package provides the estimation of the incubation period of an infectious disease}
\description{
Internal functions for the IncPeriod package.
}
\value{
No return value, called for side effects
}
